# notification-devices-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sponti API",
    "description": "Private social matching API for Sponti.\n\nContract notes for mobile clients:\n- DTO field names are stable within v1.\n- Errors use RFC 9457 ProblemDetail JSON plus stable `code` and `path` properties.\n- Clients should branch on `code`, not on localized or human-readable `detail`.\n",
    "contact": {
      "name": "MGR Tech"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/notifications/devices": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Register notification device token",
        "description": "Mobile app registers or refreshes the current push notification token for the authenticated user.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid request. Stable codes include VALIDATION_ERROR, MALFORMED_JSON, and INVALID_REQUEST.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "detail": "Invalid request",
                      "type": "about:blank",
                      "status": 400,
                      "code": "VALIDATION_ERROR",
                      "path": "/api/v1/notifications/devices",
                      "title": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. Mobile clients should show a generic retry message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR",
                    "value": {
                      "detail": "Internal server error",
                      "type": "about:blank",
                      "status": 500,
                      "code": "INTERNAL_SERVER_ERROR",
                      "path": "/api/v1/notifications/devices",
                      "title": "Internal Server Error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "AUTHENTICATION_REQUIRED": {
                    "summary": "AUTHENTICATION_REQUIRED",
                    "value": {
                      "detail": "Authentication required",
                      "type": "about:blank",
                      "status": 401,
                      "code": "AUTHENTICATION_REQUIRED",
                      "path": "/api/v1/notifications/devices",
                      "title": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete notification device token",
        "description": "Mobile app deletes the current push notification token during logout or token invalidation.",
        "operationId": "delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Invalid request. Stable codes include VALIDATION_ERROR, MALFORMED_JSON, and INVALID_REQUEST.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "detail": "Invalid request",
                      "type": "about:blank",
                      "status": 400,
                      "code": "VALIDATION_ERROR",
                      "path": "/api/v1/notifications/devices",
                      "title": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. Mobile clients should show a generic retry message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "summary": "INTERNAL_SERVER_ERROR",
                    "value": {
                      "detail": "Internal server error",
                      "type": "about:blank",
                      "status": 500,
                      "code": "INTERNAL_SERVER_ERROR",
                      "path": "/api/v1/notifications/devices",
                      "title": "Internal Server Error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing, expired, or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "AUTHENTICATION_REQUIRED": {
                    "summary": "AUTHENTICATION_REQUIRED",
                    "value": {
                      "detail": "Authentication required",
                      "type": "about:blank",
                      "status": 401,
                      "code": "AUTHENTICATION_REQUIRED",
                      "path": "/api/v1/notifications/devices",
                      "title": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "RegisterDeviceTokenRequest": {
        "type": "object",
        "description": "Register notification device token request payload",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ],
            "example": "ANDROID"
          },
          "token": {
            "type": "string",
            "example": "token",
            "minLength": 1
          },
          "deviceId": {
            "type": "string",
            "example": "device-1",
            "maxLength": 120,
            "minLength": 0
          },
          "appVersion": {
            "type": "string",
            "example": "1.0",
            "maxLength": 60,
            "minLength": 0
          }
        },
        "required": [
          "platform",
          "token"
        ]
      },
...
      "DeleteDeviceTokenRequest": {
        "type": "object",
        "description": "Delete notification device token request payload",
        "properties": {
          "token": {
            "type": "string",
            "example": "token",
            "minLength": 1
          }
        },
        "required": [
          "token"
        ]
      },
      "ApiError": {
        "type": "object",
        "description": "Stable error response. `code` is the mobile-safe discriminator.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Invalid request"
          },
          "code": {
            "type": "string",
            "description": "Stable API error code. Mobile clients should branch on this value.",
            "enum": [
              "INVALID_REQUEST",
              "MALFORMED_JSON",
              "VALIDATION_ERROR",
              "RESOURCE_NOT_FOUND",
              "INTERNAL_SERVER_ERROR",
              "AUTHENTICATION_REQUIRED",
              "UNSUPPORTED_AUTHENTICATION",
              "BAD_CREDENTIALS",
              "INVALID_REFRESH_TOKEN",
              "PHONE_NUMBER_ALREADY_USED",
              "USER_NOT_FOUND",
              "USER_NOT_FOUND",
              "AVAILABILITY_RULE_NOT_FOUND",
              "INVALID_AVAILABILITY_RULE_TIME_RANGE",
              "INVALID_AVAILABILITY_OVERRIDE_TIME_RANGE",
              "CONTACT_SELF_OPERATION_NOT_ALLOWED",
              "CONTACT_INVITATION_ALREADY_EXISTS",
              "CONTACT_ALREADY_EXISTS",
              "CONTACT_INVITATION_NOT_FOUND",
              "CONTACT_NOT_FOUND",
              "CONTACT_BLOCKED",
              "MATCHING_DISABLED",
              "PHONE_NUMBER_REQUIRED_FOR_MATCH_ACCEPTANCE",
              "USER_DOES_NOT_BELONG_TO_MATCH",
              "CHANNEL_NOT_ALLOWED",
              "AVAILABILITY_OVERLAP_NOT_FOUND",
              "MATCH_SCORE_BELOW_THRESHOLD",
              "MATCH_ALREADY_EXISTS",
              "MATCH_PROPOSAL_EXPIRED",
              "ACCEPTED_CONTACT_NOT_FOUND",
              "MATCH_NOT_FOUND",
              "MATCH_STATE_CONFLICT"
            ],
            "example": "VALIDATION_ERROR"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}